Nova controller ironic section fix for AIO
It appeared that on all-in-one environments nova.conf
file was not populated with ironic compute section details
since controller and compute states are running on
one single node. The template nova-controller.conf.Debian
was changed by adding jinja items for ironic-related items.
As a consequence ironic compute details should be added
to nova:controller pillar inside the model.

Change-Id: Ie7ce912d83193816a184debf4d4cc3764f1e1469